Projects: Move watcher to its utils
