+++ spyder/plugins/projects/utils/watcher.py
